% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alman_bib_A3a_GlobalHeatmap.R
\name{alman_bib_A3a_GlobalHeatmap}
\alias{alman_bib_A3a_GlobalHeatmap}
\title{Global Productivity Heatmap}
\usage{
alman_bib_A3a_GlobalHeatmap(
  data,
  affil_col = "Affiliations",
  low_col = "#d6eaf8",
  high_col = "#2980b9",
  top_n_labels = 5
)
}
\arguments{
\item{data}{A data frame containing the bibliographic data.}

\item{affil_col}{String. Column name for Affiliations (default: "Affiliations").}

\item{low_col}{Color for low productivity countries (default: "#d6eaf8").}

\item{high_col}{Color for high productivity countries (default: "#2980b9").}

\item{top_n_labels}{Integer. Jumlah top negara produktif yang ingin ditampilkan pada anotasi peta (default: 5).}
}
\description{
This function analyzes the geographical distribution of publications.
It automatically extracts country names from the "Affiliations" column,
standardizes them, and projects the publication frequency onto a world map.
}
\examples{
\dontrun{
# 1. Standard Heatmap (Menampilkan Top 5 negara)
alman_bib_A3a_GlobalHeatmap(df_herbal_snakebite)

# 2. Menampilkan Top 10 negara dengan tema warna "Fire" (Merah/Oranye)
alman_bib_A3a_GlobalHeatmap(
  data = df_herbal_snakebite,
  low_col = "#f6ddcc",
  high_col = "#e74c3c",
  top_n_labels = 10)
}
}
