% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_herbal_dengue}
\alias{df_herbal_dengue}
\title{Dataset from Berlianda Nur Sabilla et all (Herbal Medicine field in Tropical Disease Dengue)}
\format{
Sebuah data frame dengan 156 baris dan 46 kolom:
\describe{
\item{Authors}{Nama penulis artikel, dipisahkan dengan titik koma (;)}
\item{Title}{Judul artikel publikasi}
\item{Year}{Tahun artikel diterbitkan}
\item{Cited by}{Jumlah total sitasi yang diterima artikel tersebut}
\item{Affiliations}{Afiliasi institusi dan negara asal penulis}
\item{Keywords}{Kata kunci yang digunakan dalam artikel}
}
}
\source{
Data diekstrak dari \url{https://www.scopus.com/}
}
\usage{
df_herbal_dengue
}
\description{
This dataset extracted from "Dengue fever and herbal medicine in Indonesia: Trends, collaborations, and research impact (1985-2023)"
\url{https://doi.org/10.1051/bioconf/202515403003}
}
\details{
Abstract: The research aims to determine the extent to which herbal research in dengue fever through a bibliometric analysis in Scopus databases from 1985 to 2023. Using RStudio with the Bibliometrix package, this study scrutinizes publication patterns, identifies leading institutions, highlights influential authors, and examines keyword trends. Over this period, 122 documents were published across 88 different sources, revealing an annual growth rate of 8.06\%. This growth suggests a rising interest in the intersection of herbal medicine and dengue fever, indicating that the research community is increasingly recognizing the potential of herbal treatments. The collaborative nature of this research is underscored by the fact that the average number of co-authors per document is 5.79. Additionally, 21.31\% of these publications involve international co-authorship, reflecting a significant level of global collaboration and a shared interest in addressing this persistent health issue. Leading Indonesian institutions, such as Universitas Airlangga and Gajah Mada University, have been at the forefront of this research, making substantial contributions to the scholarly output in this field. Prominent authors, including Sucipto TH and Yohan B, have emerged as key figures. A keyword co-occurrence analysis identified “Aedes aegypti,” “animal model,” and “plant extract” as central themes in the research, which highlights a strong focus on preclinical studies and the chemical analysis of plant-based treatments.
}
\keyword{datasets}
