% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_herbal_filaria}
\alias{df_herbal_filaria}
\title{Dataset from Arista Elda Monica et all (Herbal Medicine field in Tropical Disease Filariasis)}
\format{
Sebuah data frame dengan 156 baris dan 46 kolom:
\describe{
\item{Authors}{Nama penulis artikel, dipisahkan dengan titik koma (;)}
\item{Title}{Judul artikel publikasi}
\item{Year}{Tahun artikel diterbitkan}
\item{Cited by}{Jumlah total sitasi yang diterima artikel tersebut}
\item{Affiliations}{Afiliasi institusi dan negara asal penulis}
\item{Keywords}{Kata kunci yang digunakan dalam artikel}
}
}
\source{
Data diekstrak dari \url{https://www.scopus.com/}
}
\usage{
df_herbal_filaria
}
\description{
This dataset extracted from "Herbal and filariasis in Indonesia: A citation network analysis"
\url{https://doi.org/10.1051/bioconf/202515403002}
}
\details{
Abstract: This study examines the role of herbal medicine, deeply rooted in Indonesia’s rich culture, as a potential solution for filariasis. The primary goal is to analyze the extent of herbal research related to filariasis through a citation network analysis of studies published in Scopus databases from 1980 to 2023. The study uses RStudio and the Bibliometrix package to assess publication trends, key contributing institutions, influential authors, and keyword patterns. One hundred eighteen documents have been published in 90 sources, with an annual growth rate of 5.95\%, indicating a growing interest in herbal filariasis research. The average number of coauthors per document is 4.64, and 34.75\% of the publications involve international collaboration, underscoring significant global involvement. Universitas Negeri Surabaya, Universitas Indonesia, and Universitas Gadjah Mada are leading the research efforts in this field, contributing notably to the body of knowledge. Prominent authors like Maizels RM, Djati MS, and Yazdanbach SH play a key role in shaping the academic discourse. The analysis of keyword co-occurrence reveals that “brugia malayi,” “controlled study,” and “ extract” are central themes, reflecting a strong focus on preclinical research and the extraction procedures of plant-based treatments.
}
\keyword{datasets}
