% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_herbal_leptospirosis}
\alias{df_herbal_leptospirosis}
\title{Dataset for bibliometric practice coming from Ade Dian Shah Putri et all (Herbal Medicine field in Tropical Disease Leptospirosis)}
\format{
Sebuah data frame dengan 156 baris dan 46 kolom:
\describe{
\item{Authors}{Nama penulis artikel, dipisahkan dengan titik koma (;)}
\item{Title}{Judul artikel publikasi}
\item{Year}{Tahun artikel diterbitkan}
\item{Cited by}{Jumlah total sitasi yang diterima artikel tersebut}
\item{Affiliations}{Afiliasi institusi dan negara asal penulis}
\item{Keywords}{Kata kunci yang digunakan dalam artikel}
}
}
\source{
Data diekstrak dari \url{https://www.scopus.com/}
}
\usage{
df_herbal_leptospirosis
}
\description{
This dataset extracted from "Herbal medicine and leptospirosis in Southeast Asia: A comprehensive bibliometric analysis (1973-2023)"
\url{https://doi.org/10.1051/bioconf/202515404001}
}
\details{
Abstract: This study provides a bibliometric analysis of research on herbal medicine and leptospirosis in Southeast Asia from 1973 to 2023. The investigation identified 3,043 documents through a systematic search in the Scopus database using keywords. One hundred forty-seven papers were published across 100 different sources, revealing an annual growth rate of 6.07\%, indicating a growing interest in the intersection of herbal medicine and leptospirosis. The study shows significant collaboration, with 935 authors contributing to the body of research and an average of 7.55 co-authors per document. Notably, 40.14\% of the publications involve international co-authorship, reflecting the global relevance and collaborative efforts in addressing leptospirosis through herbal remedies. Keyword analysis highlights “leptospirosis,” “leptospira,” and “DNA extraction” as central themes, demonstrating a focus on the genetic and diagnostic aspects of the disease alongside the exploration of plant-based treatments. The research also emphasizes the role of preclinical studies and the chemical analysis of herbal remedies for leptospirosis. The study identifies key contributors to the field, with prolific authors such as Chee HY, Sekawi Z, and Patarakul K leading the research efforts. Malaysia, Thailand, and Japan are highlighted as the most productive countries, significantly contributing to this domain’s research output and citation impact.
}
\keyword{datasets}
