% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_herbal_snakebite}
\alias{df_herbal_snakebite}
\title{Dataset for bibliometric practice coming from Ilham Jaluludin et all (Herbal Medicine field in Tropical Disease Snake bite)}
\format{
Sebuah data frame dengan 156 baris dan 46 kolom:
\describe{
\item{Authors}{Nama penulis artikel, dipisahkan dengan titik koma (;)}
\item{Title}{Judul artikel publikasi}
\item{Year}{Tahun artikel diterbitkan}
\item{Cited by}{Jumlah total sitasi yang diterima artikel tersebut}
\item{Affiliations}{Afiliasi institusi dan negara asal penulis}
\item{Keywords}{Kata kunci yang digunakan dalam artikel}
}
}
\source{
Data diekstrak dari \url{https://www.scopus.com/}
}
\usage{
df_herbal_snakebite
}
\description{
This dataset extracted from "Herbal medicine and snakebite research in Southeast Asia: A bibliometric analysis (1974-2024)"
\url{https://doi.org/10.1051/bioconf/202515403005}
}
\details{
Abstract: This study presents a bibliometric analysis of research on herbal medicine and snakebite in Southeast Asia, focusing on publications from 1974 to 2023. The investigation involved 205 documents selected using specific keywords in the Scopus database. After applying inclusion criteria such as language and regional focus, the final selection yielded 202 documents across 145 sources. The analysis reveals an annual growth rate of 4.25\%, indicating a steady increase in research interest over the past five decades. With contributions from 1,049 authors and an average of 6.08 coauthors per document, the field exhibits a high level of collaboration. Notably, nearly half of the publications (49.5\%) involve international coauthorship, underscoring the global significance of the research in this domain. Keyword analysis identified “snakebite,” “venom,” and “medicinal plants” as central themes. The research also emphasizes preclinical studies, which are crucial for understanding the efficacy and safety of herbal remedies for snakebite treatment. The study further identifies key contributors to the field, with prolific authors such as Tan NH, Pithayanukul P, and Gopalakrishnakone P leading the scholarly discourse. Malaysia, Thailand, and Singapore are highlighted as the most productive countries, contributing significantly to the research output and citation impact.
}
\keyword{datasets}
