% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_herbal_tb}
\alias{df_herbal_tb}
\title{Dataset from Annisa Nur Azizah et all (Herbal Medicine field in Tropical Disease Tuberculosis)}
\format{
Sebuah data frame dengan 156 baris dan 46 kolom:
\describe{
\item{Authors}{Nama penulis artikel, dipisahkan dengan titik koma (;)}
\item{Title}{Judul artikel publikasi}
\item{Year}{Tahun artikel diterbitkan}
\item{Cited by}{Jumlah total sitasi yang diterima artikel tersebut}
\item{Affiliations}{Afiliasi institusi dan negara asal penulis}
\item{Keywords}{Kata kunci yang digunakan dalam artikel}
}
}
\source{
Data diekstrak dari \url{https://www.scopus.com/}
}
\usage{
df_herbal_tb
}
\description{
This dataset extracted from "Herbal and tuberculosis in Indonesia: Bibliometric analysis"
\url{https://doi.org/10.1051/bioconf/202515403001}
}
\details{
Abstract: This study explores the herbal status, grounded in Indonesia’s extensive tradition of herbal medicine, to address tuberculosis. The primary aim is to elucidate the extent of herbal research in tuberculosis through a bibliometric analysis of relevant research published in Scopus databases between 1996 and 2023. The study used RStudio, complemented by the Bibliometrix package, to assess publication patterns, contributing institutions, influential authors, and keyword trends. Two hundred-one documents have been published across 138 different sources, with an annual growth rate of 13.83\%, suggesting possibilities of growing interest in herbal tuberculosis research. The average number of co-authors per document is 5.65, and 21.89\% of the publications involve international co-authorship, highlighting a significant level of global collaboration. Universitas Padjajaran and Airlanga are at the forefront of research in this domain, contributing significantly to the scholarly output. Noteworthy contributions by prolific authors such as Massi MN and Mertaniasih underscore the critical nodes of academic influence. Keyword co-occurrence analysis identified “mycobacterium tuberculosis,” “animal experiment,” and “plant extract” as central themes, highlighting a predominant focus on preclinical research and the chemical analysis of plant-based treatments.
}
\keyword{datasets}
