% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_leadership}
\alias{df_leadership}
\title{Contoh Data Bibliometrik untuk Analisis ARFAN kasus manajerial kepemimpinan}
\format{
Sebuah data frame dengan 500 baris dan 10 kolom:
\describe{
\item{Authors}{Nama penulis artikel, dipisahkan dengan titik koma (;)}
\item{Title}{Judul artikel publikasi}
\item{Year}{Tahun artikel diterbitkan}
\item{Cited by}{Jumlah total sitasi yang diterima artikel tersebut}
\item{Affiliations}{Afiliasi institusi dan negara asal penulis}
\item{Keywords}{Kata kunci yang digunakan dalam artikel}
}
}
\source{
Data diekstrak dari \url{https://www.scopus.com/} sebagai sampel anonim.
}
\usage{
df_leadership
}
\description{
Dataset ini berisi contoh data ekspor dari Scopus yang digunakan
untuk mendemonstrasikan fungsi-fungsi di dalam package ARFAN,
seperti pembuatan Executive Summary Dashboard.
}
\keyword{datasets}
