% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_patientsafety}
\alias{df_patientsafety}
\title{Dataset for bibliometric practice coming from Hospital Management field in patient safety domain.}
\format{
Sebuah data frame dengan 500 baris dan 10 kolom:
\describe{
\item{Authors}{Nama penulis artikel, dipisahkan dengan titik koma (;)}
\item{Title}{Judul artikel publikasi}
\item{Year}{Tahun artikel diterbitkan}
\item{Cited by}{Jumlah total sitasi yang diterima artikel tersebut}
\item{Affiliations}{Afiliasi institusi dan negara asal penulis}
\item{Keywords}{Kata kunci yang digunakan dalam artikel}
}
}
\source{
Data diekstrak dari \url{https://www.scopus.com/}
}
\usage{
df_patientsafety
}
\description{
This dataset extracted from "EVOLUTION OF ARTIFICIAL INTELLIGENCE IN PATIENT SAFETY ACROSS SOUTHEAST ASIA: A BIBLIOMETRIC ANALYSIS"
\url{https://doi.org/10.53806/iamsph.v7i1.1447}
}
\details{
Abstract: This study evaluated the evolution of artificial intelligence (AI) and machine learning in patient safety across Southeast Asia (SEA) to identify regional research dynamics and emerging frontiers. A quantitative bibliometric analysis was conducted using 262 primary documents retrieved from the Scopus database (English-language; published up to December 31, 2025). Document selection was guided by an adaptation of the PRISMA-ScR framework. Key indicators included the Mann-Kendall trend test, Pettitt’s change-point test, and thematic cooccurrence network mapping. Statistical analysis revealed a significant structural shift in 2015 (p < 0.01), marking exponential publication growth and extensive international collaboration (62.2\% of documents featuring multi-country co-authorship). Results highlighted a divergence in regional strategies; high-income nations produced high-impact clinical algorithms, while emerging economies prioritized capacity building for resource-constrained systems. Thematic mapping demonstrated a major transitional shift in the literature from traditional, image-based neural networks for diagnostic accuracy toward natural language processing and large language models aimed at addressing clinical documentation and prescribing errors. These publication dynamics map the current academic focus, highlighting the need for future implementation driven clinical validation and regionally adapted regulatory frameworks.
}
\keyword{datasets}
